chore: refactor NewTaskViewController
