+++ Challenge04/Challenge04/Tools/HapticsManager.swift
